+++ docs/changes.rst
@@ -7,6 +7,7 @@ Version 0.1.1
 To be released.
 
 - Added :class:`sider.types.Date` type.
+- Added :class:`sider.datetime.FixedOffset` tzinfo subtype.
 
 
 Version 0.1.0